+++ src/main/java/org/jolokia/docker/maven/util/WaitUtil.java
@@ -2,11 +2,11 @@ package org.jolokia.docker.maven.util;
 
 import java.io.IOException;
 import java.util.concurrent.TimeoutException;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import org.apache.http.client.config.RequestConfig;
 import org.apache.http.client.methods.CloseableHttpResponse;
-import org.apache.http.client.methods.HttpHead;
-import org.apache.http.client.methods.HttpUriRequest;
 import org.apache.http.client.methods.RequestBuilder;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClientBuilder;
@@ -25,7 +25,14 @@ public class WaitUtil {
 
     // Timeout for ping
     private static final int HTTP_PING_TIMEOUT = 500;
+
+    // Default HTTP Method to use
     public static final String DEFAULT_HTTP_METHOD = "HEAD";
+
+    // Default status codes
+    public static final int DEFAULT_MIN_STATUS = 200;
+    public static final int DEFAULT_MAX_STATUS = 399;
+
 
     private WaitUtil() {}
 
@@ -82,6 +89,7 @@ public class WaitUtil {
      */
     public static class HttpPingChecker implements WaitChecker {
 
+        private int statusMin,statusMax;
         private String url;
         private String method;
 
@@ -89,11 +97,33 @@ public class WaitUtil {
          * Ping the given URL
          *
          * @param url URL to check
-         * @param method
+         * @param method HTTP method to use
+         * @param status status code to check
          */
-        public HttpPingChecker(String url, String method) {
+        public HttpPingChecker(String url, String method, String status) {
             this.url = url;
             this.method = method;
+
+            if (method == null) {
+                this.method = DEFAULT_HTTP_METHOD;
+            }
+
+            if (status == null) {
+                statusMin = DEFAULT_MIN_STATUS;
+                statusMax = DEFAULT_MAX_STATUS;
+            } else {
+                Matcher matcher = Pattern.compile("^(\\d+)\\s*\\.\\.+\\s*(\\d+)$").matcher(status);
+                if (matcher.matches()) {
+                    statusMin = Integer.parseInt(matcher.group(1));
+                    statusMax = Integer.parseInt(matcher.group(2));
+                } else {
+                    statusMin = statusMax = Integer.parseInt(status);
+                }
+            }
+        }
+
+        public HttpPingChecker(String waitUrl) {
+            this(waitUrl,null,null);
         }
 
         @Override
@@ -116,13 +146,13 @@ public class WaitUtil {
                     .setDefaultRequestConfig(requestConfig)
                     .build();
             try {
-                if (method == null) {
-                    method = DEFAULT_HTTP_METHOD;
-                }
-                CloseableHttpResponse response = httpClient.execute(RequestBuilder.create(method).setUri(url).build());
+                CloseableHttpResponse response = httpClient.execute(RequestBuilder.create(method.toUpperCase()).setUri(url).build());
                 try {
                     int responseCode = response.getStatusLine().getStatusCode();
-                    return (responseCode >= 200 && responseCode <= 399);
+                    if (responseCode == 501) {
+                        throw new IllegalArgumentException("Invalid or not supported HTTP method '" + method.toUpperCase() + "' for checking " + url);
+                    }
+                    return (responseCode >= statusMin && responseCode <= statusMax);
                 } finally {
                     response.close();
                 }