+++ doc/changelog.md
@@ -10,7 +10,8 @@
   - Add new configuration parameter 'maxConnections' to allow to specify the number of parallel 
     connections to the Docker Host. Default: 100 (#254)
   - Allow multiple containers of the same image to be linked (#182)
-
+  - HTTP method and status code can be specified when waiting on an HTTP URL (#258)
+   
 * **0.13.2**
   - "run" directives can be added to the Dockerfile (#191)
   - Support user information in wait URL (#211)